requirements.txt
# Core LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# OpenAI
openai>=1.0.0

# Web Search
duckduckgo-search>=6.0.0

# Web UI
gradio>=4.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Timezone
pytz>=2024.1
